extended NodeRouter from Router
