+++ packages/node/src/api-router.ts
@@ -1,5 +1,3 @@
-import { Router } from "rpc-server";
-
 import {
   addChannelController,
   depositEventController,
@@ -31,6 +29,8 @@ import {
   UpdateStateController,
   WithdrawController
 } from "./methods";
+import { RequestHandler } from "./request-handler";
+import NodeRouter from "./rpc-router";
 import { NODE_EVENTS } from "./types";
 
 const controllers = [
@@ -84,7 +84,8 @@ export const methodNameToImplementation = controllers.reduce(
   {}
 );
 
-export const createRpcRouter = () => new Router({ controllers });
+export const createRpcRouter = (requestHandler: RequestHandler) =>
+  new NodeRouter({ controllers, requestHandler });
 
 export const eventNameToImplementation = {
   [NODE_EVENTS.CREATE_CHANNEL]: addChannelController,
